requirements.txt
ccxt==4.1.51
pandas==2.1.4
numpy==1.24.3
python-dotenv==1.0.0
matplotlib==3.4.3
seaborn==0.11.2
